requirements/01_requirements.txt
click==7.1.2
biopython==1.77